requirements.txt
langchain==1.2.13
langchain-community==0.4.1
langchain-ollama=1.0.1
chromadb==1.5.5
sentence-transformers==5.3.0
pymupdf==1.27.2.2
google-genai==0.6.0